AI/search.py
import math
import os

from tensorflow.keras.preprocessing import image
from tensorflow.keras.applications.vgg16 import VGG16, preprocess_input
from tensorflow.keras.models import  Model

from PIL import Image
import pickle
import numpy as np

# Ham tao model
def get_extract_model():
    vgg16_model = VGG16(weights="imagenet")
    extract_model = Model(inputs=vgg16_model.inputs, outputs = vgg16_model.get_layer("fc1").output)
    return extract_model

# Ham tien xu ly, chuyen doi hinh anh thanh tensor
def image_preprocess(img):
    img = img.resize((224,224))
    img = img.convert("RGB")
    x = image.img_to_array(img)
    x = np.expand_dims(x, axis=0)
    x = preprocess_input(x)
    return x

def extract_vector(model, image_path):
    print("Xu ly : ", image_path)
    img = Image.open(image_path)
    img_tensor = image_preprocess(img)

    # Trich dac trung
    vector = model.predict(img_tensor)[0]
    # Chuan hoa vector = chia chia L2 norm (tu google search)
    vector = vector / np.linalg.norm(vector)
    return vector

def search(arrayVectors, image_path):
    model = get_extract_model()
    search_vector = extract_vector(model, image_path) 
    # type search_vector: <class 'numpy.ndarray'>

    vectors = [(np.frombuffer(x.vector, dtype=np.float32)) for x in arrayVectors]
    # type vectors: [<class 'bytes'>] -> [<class 'numpy.ndarray'>]

    # Tinh khoang cach tu search_vector den tat ca cac vector
    distance = np.linalg.norm(vectors - search_vector, axis=1)
   
    # Sap xep va lay ra vector so 0 la vector có khoang cach gan nhat
    ids = np.argsort(distance)[0]
    return arrayVectors[ids].name